Export Xgettext as Renderer
