+++ i18n-helpers/src/bin/mdbook-xgettext.rs
@@ -22,25 +22,12 @@
 //! of Chapter nesting.
 
 use anyhow::Context;
-use mdbook::renderer::RenderContext;
-use mdbook_i18n_helpers::xgettext::create_catalogs;
-use std::{fs, io};
+use mdbook::renderer::{RenderContext, Renderer};
+use mdbook_i18n_helpers::renderers::Xgettext;
+use std::io;
 
 fn main() -> anyhow::Result<()> {
     let ctx = RenderContext::from_json(&mut io::stdin()).context("Parsing stdin")?;
-    fs::create_dir_all(&ctx.destination)
-        .with_context(|| format!("Could not create {}", ctx.destination.display()))?;
-    let catalogs = create_catalogs(&ctx, std::fs::read_to_string).context("Extracting messages")?;
-
-    // Create a template file for each entry with the content from the respective catalog.
-    for (file_path, catalog) in catalogs {
-        let directory_path = file_path.parent().unwrap();
-        fs::create_dir_all(directory_path)
-            .with_context(|| format!("Could not create {}", directory_path.display()))?;
-
-        polib::po_file::write(&catalog, &file_path)
-            .with_context(|| format!("Writing messages to {}", file_path.display()))?;
-    }
-
-    Ok(())
+    let renderer = Xgettext {};
+    renderer.render(&ctx)
 }
