+++ i18n-helpers/src/renderers.rs
@@ -1,0 +1,2 @@
+mod xgettext;
+pub use xgettext::Xgettext;
